pyproject.toml
[project]
name = "jamly"
version = "0.1.0"
description = ""
authors = [
    {name = "sbrad93",email = "shannon.brady2@marist.edu"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "spotipy (>=2.25.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "jamly", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
